requirements.txt
fastapi==0.104.1
streamlit==1.28.1
openai==1.102.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
numpy>=1.21.0,<1.25.0
scikit-learn==1.3.2
python-multipart==0.0.6
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pandas>=1.5.0,<2.1.0
statsmodels==0.14.5